# swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diagnoxix API",
        "title": "Diagnoxix",
        "contact": {},
        "version": "1.0"
    },
    "host": "diagnoxix.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/medical_records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all medical records for the authenticated user with pagination and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "List user's medical records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Records per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "LAB_REPORT",
                            "PRESCRIPTION",
                            "IMAGING",
                            "DISCHARGE_SUMMARY",
                            "OTHER"
                        ],
                        "type": "string",
                        "description": "Filter by document type",
                        "name": "document_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Filter by date from (YYYY-MM-DD)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Filter by date to (YYYY-MM-DD)",
                        "name": "to_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of medical records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a new medical record with metadata and file attachment",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "Upload a new medical record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "123e2345-e89b-12d3-a456-426614174000",
                        "description": "Uploader ID (for diagnostic centres)",
                        "name": "uploader_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "673e9876-e89b-12d3-a456-426614174000",
                        "description": "Associated Schedule ID",
                        "name": "schedule_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Record title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "LAB_REPORT",
                            "PRESCRIPTION",
                            "DISCHARGE_SUMMARY",
                            "IMAGING",
                            "VACCINATION",
                            "ALLERGY",
                            "SURGERY",
                            "CHRONIC_CONDITION",
                            "FAMILY_HISTORY",
                            "OTHER"
                        ],
                        "type": "string",
                        "description": "Type of medical document",
                        "name": "document_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Date of the document",
                        "name": "document_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Healthcare provider name",
                        "name": "provider_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Medical specialty",
                        "name": "specialty",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Medical record file (PDF/Image)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Medical record created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                        }
                    },
                    "400": {
                        "description": "Invalid input data/file format",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "403": {
                        "description": "Insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "DUPLICATE_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.DUPLICATE_ERROR"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/medical_records/diagnostic_centre/{diagnostic_centre_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all medical records uploaded by a diagnostic centre with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "List uploaded medical records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Records per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of medical records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/medical_records/{record_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific medical record. Access limited to record owner or authorized diagnostic centre.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "Get a medical record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Medical Record ID",
                        "name": "record_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical record details",
                        "schema": {
                            "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update metadata of an existing medical record. File content cannot be updated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "Update a medical record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "987e4567-e89b-12d3-a456-426614174000",
                        "description": "Medical Record ID",
                        "name": "record_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated record details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateMedicalRecordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/medical_records/{record_id}/diagnostic_centre/{diagnostic_centre_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a medical record uploaded by a diagnostic centre",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecord"
                ],
                "summary": "Get an uploaded medical record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Medical Record ID",
                        "name": "record_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "675e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical record details",
                        "schema": {
                            "$ref": "#/definitions/handlers.MedicalRecordSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BAD_REQUEST": {
            "description": "BAD_REQUEST response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handlers.DUPLICATE_ERROR": {
            "description": "DUPLICATE_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "DUPLICATE_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Duplicate error"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 409
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "description": "UNAUTHORIZED_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.INTERNAL_SERVER_ERROR": {
            "description": "INTERNAL_SERVER_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "INTERNAL_SERVER_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Internal Server Error"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.MedicalRecordSwagger": {
            "description": "Medical record response for Swagger",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-06-26T20:00:00Z"
                },
                "diagnostic_centre_id": {
                    "type": "string",
                    "example": "dc-001"
                },
                "file_type": {
                    "type": "string",
                    "example": "PDF"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://medivue.com/records/rec-001.pdf"
                },
                "id": {
                    "type": "string",
                    "example": "rec-001"
                },
                "patient_id": {
                    "type": "string",
                    "example": "user-001"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-26T20:30:00Z"
                }
            }
        },
        "handlers.NOT_FOUND_ERROR": {
            "description": "NOT_FOUND_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "NOT_FOUND_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Data not found"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.UNAUTHORIZED_ERROR": {
            "description": "UNAUTHORIZED_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "UNAUTHORIZED_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Unauthorized to perform this operation"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}